remove unecessary hook
