+++ src/config.ts
@@ -1,6 +1,8 @@
 const isDev = import.meta.env.DEV;
 const isMobile = import.meta.env.MODE === 'app';
 
+export const APP_TITLE = 'aula';
+
 export interface RuntimeConfig {
   CENTRAL_API_URL: string;
   IS_MULTI: boolean;
@@ -36,7 +38,7 @@ export async function loadRuntimeConfig(): Promise<RuntimeConfig> {
   return config;
 }
 
-export class RuntimeConfigNotFoundError extends Error { }
+export class RuntimeConfigNotFoundError extends Error {}
 
 export function getRuntimeConfig(): RuntimeConfig {
   const config = localStorage.getItem('config');
