+++ src/hooks/instance.ts
@@ -1,7 +1,6 @@
 import { localStorageGet, localStorageSet } from '@/utils';
 import { useEffect, useState } from 'react';
 import { useLocation, useNavigate } from 'react-router-dom';
-import { useIsOnline } from './auth';
 
 export const useCodeManagement = () => {
   const location = useLocation();
@@ -21,23 +20,5 @@ export const useCodeManagement = () => {
   return {
     code,
     resetCode,
-  };
-};
-
-export const useIsOnlineState = () => {
-  const location = useLocation();
-  const [online, setOnline] = useState(true);
-
-  const checkOnlineStatus = async () => {
-    const isOnlineResult = await useIsOnline();
-    setOnline(isOnlineResult);
-  };
-
-  useEffect(() => {
-    checkOnlineStatus();
-  }, [location]);
-
-  return {
-    online,
   };
 };
